refactor/ Changed application architecture
