+++ src/components/admin/create-product-container/create-product-container.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import Button from '../../button/button'
+import Button from '../../global/button/button'
 import styles from './create-product-container.module.css'
 
 const CreateProductContainer = () => {
