+++ src/components/global/auth-title/auth-title.js
