+++ src/components/global/font-awesome-icon/fontAwesomeIcon.js
