+++ src/components/global/footer/footer.module.css
@@ -1,9 +1,6 @@
 .footer {
     display: flex;
     justify-content: center;
-    /* align-items: center; */
-    /* align-content: center; */
-    
     padding: 20px;
     position: absolute;
     bottom: 0;
