+++ src/components/global/page-layout/page-layout.module.css
@@ -1,7 +1,6 @@
 .app {
     max-width: 100%;
     margin: 0 auto;
-    /* position: relative; */
     position: relative;
     min-height: 100vh;
 }
