+++ src/components/home/second-home-section/grid-item/grid-item.module.css
