+++ src/components/products/grid-item/grid-item.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import styles from './grid-item.module.css'
-import Button from '../../button/button'
-import FontAwIcon from '../../font-awesome-icon/fontAwesomeIcon'
-import PriceBox from '../../small-utils/price-box/price-box'
+import Button from '../../global/button/button'
+import FontAwIcon from '../../global/font-awesome-icon/fontAwesomeIcon'
+import PriceBox from '../../global/price-box/price-box'
 import { faCartPlus } from '@fortawesome/free-solid-svg-icons'
 import { withRouter } from 'react-router'
 import UserContext from '../../../Context'
