+++ src/pages/cart-page/cart-page.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Cart from '../../components/cart/cart'
-import PageLayout from '../../components/page-layout/page-layout'
+import PageLayout from '../../components/global/page-layout/page-layout'
 
 const CartPage = () => {
