+++ src/pages/home-page/home-page.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import FirstHomeSection from '../../components/first-home-section/first-home-section'
-import PageLayout from '../../components/page-layout/page-layout'
-import SecondHomeSection from '../../components/second-home-section/second-home-section'
+import FirstHomeSection from '../../components/home/first-home-section/first-home-section'
+import PageLayout from '../../components/global/page-layout/page-layout'
+import SecondHomeSection from '../../components/home/second-home-section/second-home-section'
 
 const HomePage = () => {
