+++ src/pages/login-page/login-page.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import AuthTitle from '../../components/auth-title/auth-title'
+import AuthTitle from '../../components/global/auth-title/auth-title'
 import Login from '../../components/auth/login'
-import PageLayout from '../../components/page-layout/page-layout'
+import PageLayout from '../../components/global/page-layout/page-layout'
 
 const LoginPage = () => {
